backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.3
python-multipart==0.0.12
httpx==0.27.2
pillow>=11.1.0
pillow-heif>=1.1.0
